test: MenuNameTest, MenuPriceTest 추가
