+++ src/main/java/kitchenpos/domain/menu/MenuName.java
@@ -15,11 +15,11 @@ public class MenuName {
     }
 
     public MenuName(final String name) {
-        validation(name);
+        validate(name);
         this.name = name;
     }
 
-    private void validation(final String name) {
+    private void validate(final String name) {
         if (Objects.isNull(name) || name.length() > MAX_LENGTH) {
             throw new IllegalArgumentException();
         }
